backend/requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
flask-session
redis
python-dotenv
mysqlclient
flask-marshmallow
marshmallow-sqlalchemy
flask-cors
flask-mail
flask[async]
pysha3
requests
flask-jwt-extended
yagmail